requirements.txt
colorlog==3.1.4
discord.py==0.16.12